refactor: extract send_prompt_direct() into OpenAIProxy class (t046)
Move all direct OpenAI-compatible API logic out of AgentLoop into a
dedicated OpenAIProxy class in includes/Core/. The new class encapsulates:
- Request formatting (abilities → tools, history → messages)
- Direct HTTP call via wp_remote_post()
- Response parsing into SimpleAiResult
- Error handling returning WP_Error

AgentLoop.send_prompt_direct() now constructs an OpenAIProxy instance and
delegates to proxy->send(), reducing AgentLoop by ~280 lines.

PHPStan level 5: no new errors. PHPCS: clean.
